requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aiohttp-jinja2==0.8.0
aiohttp==1.0.1
async-timeout==1.0.0      # via aiohttp
chardet==2.3.0            # via aiohttp
gunicorn==19.6.0
jinja2==2.8               # via aiohttp-jinja2
MarkupSafe==0.23          # via jinja2
multidict==2.1.0          # via aiohttp
pyyaml==3.12
